bone joints vertex data should be unsigned-short according to glTF2. This change will require a recent version of glkit, as it did not support them until my patch.

+++ data/default.vert
@@ -16,7 +16,7 @@
 
 (define-vertex-groups bones
   ((:id :bones
-    :attrs ((joints :unsigned-int 4)
+    :attrs ((joints :unsigned-short 4)
             (weights :float 4)))))
 
 ;;; vertex layouts
